rename all react-navs occurrences into react-tabs-handler; Fix some typos
